Add files via upload
Description: TeleportPads is a Minecraft server plugin that allows players to create and use pressure plates as teleportation points in the game world. With this plugin, players can place a pressure plate at a specific location and assign destination coordinates to it. When a player steps on the pressure plate, they are instantly teleported to the designated location.

Command to Create Teleport Pads:

/setteleportpad <x> <y> <z>

Command Description: To use TeleportPads, simply execute the command /setteleportpad <x> <y> <z> while in-game and standing directly on top of a pressure plate where you want to create the teleport pad. Replace <x> <y> <z> with the coordinates of the destination location where you want players to be teleported when they step on the pressure plate. Once the teleport pad is created, players can use it to quickly travel around the game world.